Add check for POLLNVAL in case a socket was closed
The SocketImpl::accept for an AF_INET socket does a loop
on ppoll() until an error or a connection comes in.  If the socket was
closed, then if the returned events are not checked for POLLNVAL,
it will loop infinitely.

I've seen this situation when testing a dsm process with
an output server socket that has not connected - one can't shut
it down without a "kill -9".

ServerSocket::close(), in nidas/core/Socket.cc, does the following:
    if (_connectionThread)
        _connectionThread->interrupt(); // this interrupt method also sends a SIGUSR1
    if (_servSock) _servSock->close();

One would think the signal will be detected by ppoll before the closed
file descriptor, but not so.  Adding checks for POLLNVAL seems to cure
the situation.