Support for Uio48 digital I/O on Vortex PCM-C418
Compiles, but with only minimal, remote testing without actually
checking or setting a voltage on the pins.

Command line command (nidas/apps/uio48.cc)
    get a pin:
        uio48 /dev/uio48a pin
    set a pin:
        uio48 /dev/uio48a pin 0|1

Pins are numbered from 0, up to 23.

Setting a pin to a value 0, connects it to ground, where it can sink
current.

Setting a pin to a value 1, changes it to a high-impedence, floating,
state where it can be used as an input. It does not set it to a plus
voltage where it could source current.

So one must set a pin to 1 in order for it to be used as an input.  That
is the default state after a reboot.

As described above, Nidas inverts the logic as descibed in the
manual for the WinSystems PCM-UIO48B:

    Writing a 0 to a bit position causes the corresponding output pin to go to a
    high-impedance state (pulled high by external 10 KΩ resistors). This allows it to be used as an
    input. When used in the input mode, a read reflects the inverted state of the I/O pin, such that
    a high on the pin will read as a 0 in the register. Writing a 1 to a bit position causes that output
    pin to sink current (up to 12 mA), effectively pulling it low.

Also included in this commit is a Uio48Sensor, which can poll the
current settings of the pins.  It remains to be seen how useful it is.
The process() method converts the binary value of the 24 pins to a
single precision float, which has an effective 24 bits of mantissa, so we
shouldn't lose any information.