require eol_scons 4.2
nidas requires install destinations to be interpolated when created and
not when installed, which eol_scons 4.2 breaks because the destination
is passed on as a string.  this can be worked around by
disabling the install alias, which is only available in eol_scons 4.2.
So start using other eol_scons features, like gitinfo=off,
versionfiles alias, and PrintProgress().
Remove remnants of BUILD_RAF.