simplify EndianConverter header and memory usage
the EndianConverter subclass instances were leaking, but it turns out all the
dynamic memory access and static mutex can be removed.  The EndianConverter
virtual interface is all that is used, so the subclass declarations do not
need to be in the header, let alone the definitions.  Also, the private static
members can be moved the implementation module rather than requiring the
header to depend on ThreadSupport and other headers.  The host endianness can
be a function-scope static, likewise for the two converter implementations, so
that makes the mutex unecessary also.  Since the subclass converters are
defined in the source module, all the inline implementation functions can be
moved there also.