move ftdi interface implementations out of headers
the GpioIF and PowerCtrlIf method definitions do not need to be in the header.
putting them in the source module reduces the dependencies of the header, and
it avoids issues like duplicating definitions of statics.  (ie, usually it is
wrong to define static const char* SYMBOL in a header, because that means
every source module which includes that header also includes memory for that
symbol.)

also try to simplify the enum conversions to string using a map instead of
long switch statements.

we are especially trying to remove gratuitous includes of ftdi.h, since no
users of the sensor power and transceiver control interfaces need to know
about that implementation detail.