allow multiple prompts in a sensor without samples
rather than only allowing multiple prompts to be added to a sensor inside
sample tags, allow multiple prompts inside the sensor, so there does not need
to be variables associated with every prompt.  remove the single prompt
instance inside sensor and consolidate on just the prompt list, where the list
always has a "primary prompt" that corresponds to the setPrompt()/getPrompt()
methods.  have some fun with auto range iterators to simplify the handling of
prompters.  all the prompts will always be part of the sensor, even the
invalid ones, so ignore the invalid prompts when creating the prompters.