remove ftdi and libusb build flags
These were dangling references from the DSM3 development, which has since
moved entirely to the buster branch, so they are not yet needed on master.
This was the wrong place to put them in the first place; I think they have
already been moved to the library SConscripts on the branch.