load netcdf classes dynamically into prep
Make addSampleTag() a virtual method of IOChannel which only NetcdfRPCChannel
implements, so prep can add sample tags to it through the IOChannel interface.
Configure NetcdfRPCChannel with an xml doc instead of direct method calls, so
it can be loaded and allocated dynamically through the IOChannel interface.
Share the code to parse an xml string into a doc in the XMLParser class.
Replace namespace pollution in NetcdfRPCChannel.h with type aliases.