simplify kernel module builds
consolidate all the similar scons setup for each kernel module into one
pseudo-builder function, NidasKernelModule().

since only one variant is being built, the kernel module builders and
construction variables only need to be setup once in one environment, and the
environment does not need to be cloned for each module.

the code which tried to clean modules from the install location even when not
being built was not working, so remove it.  the only targets which can be
cleaned from the install are the ones which are enabled to be built and
installed.

SConstruct sets VARIANT_DIR in the variant environment.  nidas/SConscript uses
this to create all the Configure artifacts under the variant dir, rather than
trying to name them all uniquely.  nidas/linux/SConscript uses it to create
linux/Revision.h in the right place and also make it available as a
SymbolModules() target.

set the variant build dir as the default target.

since kernel modules require a variant build with duplicate=1, the top-level
variant now uses duplicate=1, and that is the only variant sconscript call.
the duplicates are made with hard links, so it seems reasonable to simplify
the build by using only one variant that works for all the SConscripts.