accept BUILD or command-line target for arch-mach
instead of ARCH and MACH being scons variables, use a single BUILD variable
which can be matched against supported targets to set ARCH and MACH in the
environment.  if a command-line target matches a build target, then that
takes precedence.  BUILDS and MACHS are checked for backwards compatibility.

update the help text accordingly, and also play with a way to provide brief
variable help, unless -h --help-all is passed.  scons -Q -h is now a much more
readable summary of the most used settings.