cache device interface implementations in the base
The HardwareInterface base class can cache the interface pointers which the
subclass implementation creates.  Protected methods like
createOutputInterface() allow the base class to take care of the caching if
the subclass wants it.

Move the standard devices into a namespace instead of a struct, so they can be
used unqualified by using the namespace.

Fill in the ftdi implementation by consolidating all ftdi interaction into one
class which the interfaces use to read and write the specified bits.

Change SerialPortInterface to get and set all the port config settings in
single calls.

The device description and the interface caches are handled by a new class
HardwareDeviceImpl, meant to extend the HardwareDevice with
implementation-specific state without exposing it outside the backend.
Includes the device description, so HardwareDevice is simpler and the
implementation can easily add devices or modify descriptions of standard
devices.