implement ftdi context sharing
When a ftdi device is created, look for an existing ftdi device with the same
interface open and share that context instead.  Synchronize access to ftdi
interface and hardware.  Allow the device locations to be differentiated by
interface type, so no interface implementation needs to shift their bits,
FTDI_Device takes care of it.  Simplify serial port config interface so that
port type and termination are always written and read together, it's not worth
trying to allow them to be set separately.

This interface now works to set and get all DSM3 GPIO lines, but there are
bugs to work out in managing the mutex and interface lifetimes.