limit hardware interface lifetime with shared ptr
The current hardware interface implementation is returned as a shared pointer,
and it's up to the caller to scope the shared pointers, whether returned
directly from HardwareInterface::getHardwareInterface() or from a
HardwareDevice.  This allows implementation devices to be closed as soon as
the interface is out of scope.  Change the device cache to a std::list, and
synchronize the HardwareInterface device cache with a mutex.

Simplify pio usage and argument handling, and make the output less verbose.