remove power interface from DSMSensor
The DSMSensor interface did not need to be polluted with so many methods
related to hardware power control for serial port sensors.  Access to the
power control for a device now goes through the device name lookup.  Port
SerialSensor hardware calls to HardwareInterface.

Hardware implementation now defaults to ftdi, without actually checking if
it's supported on the current hardware, since that's current behavior anyway.

Map ttyDSM devices names to hardware ports in HardwareDevice::lookupDevice().

Simplify access to devices() and make it more consistent: devices() is
generated from the device implementations cached in the HardwareInterface, and
the returned devices already have a reference to the HardwareInterface which
generated them.  Add HardwareDevice::lookupDevice() to get a device bound to
the current hardware interface without having to first get the interface.

Make OutputState class with convenience methods toString() and fromString().