start simplifying SerialPortIODevice
Remove unused constructors.  PortConfig applied member was not actually tested
anywhere, only set, and it doesn't make sense as part of the PortConfig itself
anyway, so remove it and all signatures which set it.  In particular, remove
from autoconfig sensors the static const set to false just to pass as the
default applied value to constructors.