restore CSAT3 dataMode() to fix serial_sensor test
CSAT3_Sonic with autoconfig disabled no longer sent the D or D& command to the
sonic to start measurements, and besides being the pre-autoconfig behavior,
sensor_sim depends on that to start writing samples for the serial_sensor
test.  So restore dataMode() and use that in exitConfigMode() if autoconfig
enabled or else call it in open() as before if autoconfig is disabled.

The test results still change because there is no extra sample from the
querySonic() call that happened in the pre-autoconfig code, although it's
arguable that should be restored also, so that behavior stays the same as
before when autoconfig is disabled.