simplify autoconfig sensor construction
Remove lots of unused constructor signatures and members in autoconfig
classes.  Refactor port config management in the hopes of clarifying how
multiple port configs are assembled and then used.

The notions of "default" and "desired" and "working" configs confused the api,
so those have been renamed or removed.  A SerialSensor is now just a list of
available configs, nominally in order of preference.  Typical configs can be
hardcoded in the sensor class, but specific configs from the xml take
precedence.

In the simplest case, a sensor has one available port config, the one
specified in the xml sensor element, and that is the port config which will be
used.

In the autoconfig case, where a senor knows how to test if a port config is
working, it can cycle through the available port configs looking for one which
works.  Instead of all the static const all-caps variables and the three
separate lists of port types, bauds, and WordSpec, there is a simple vector of
PortConfig instances and an api to modify it.  This vastly simplifies the code
which searches for a working port config.

Rather than duplicate code in buildIODevice between CharacterSensor and
SerialSensor, the CharacterSensor implementation just returns null if the
device name does not have a recognized prefix, indicating to SerialSensor that
it should create a SerialPortIODevice.