Updates to usbwtod
These were motivated by what looked like a problem interfacing
probes on a Vortex (i686).  It looks like that was a signal interference
issue, not software, but as part of that investigation, this code was
given a review.

The most substantial change is related to USE_DMA_COHERENT_FOR_IMG_URBS.
If it is defined, then usb_alloc_coherent() is used instead of kmalloc()
to allocate space for the received image buffers, and URB_NO_TRANSFER_DMA_MAP
is set in the URB tranfer flags. This was the old behavior before this
change, which has been used up to now on a Vulcan (ARM).

https://www.kernel.org/doc/html/latest/driver-api/usb/dma.html indicates that
coherent memory and URB_NO_TRANSFER_DMA_MAP is appropriate when doing
lots of small transfers from the same buffer, and that most drivers
should not use it. Therefore USE_DMA_COHERENT_FOR_IMG_URBS is not
defined currently. We'll see how things work in testing.

Also cleaned up the logic and integer math related to throttle rates.
I believe the previous code would not support throttle rates > 150.
Added a bunch of comments, warnings about the maximum throttle rates and
improved some LOG messages.