Error handling cleanup
The error handling for any of the three types of URBs is identical:
set dev->errorStatus and that error will be returned to the user on
the next poll or read. So, the switches on the urb->status can
be moved to a function used by all the urb callbacks.

Call wake_up_interruptible() whenever dev->errorStatus is set.