copy pi3 buster build scripts from scons-build-single-target
scons-build-single-target includes the container-builds branch which was
being used for pi3 buster container builds, so this commit copies the
necessary files from that branch to the buster branch so this branch can
build itself for pi3.

the rest of the scripts directory will still need to be merged later, but
they are not merged here because they are tied to changes to the scons
build targets.