fix leak, add merge, fix items set in constructor
fix leak from wrong use of unique_ptr::release; add a merge() method so
metadata can be filled in without replacing other metadata.  for metadata
interfaces which assign values in the constructor, that adds to the metadata
store before it can be bound to a different metadata instance, so those
settings have to merged into the new binding to be preserved.