Move logic out of Makefile and into SCons
Since the kmake tool already determines the linux source directory, use that
to derive the kernel release name and the install path for linux modules, then
install the modules directly to that path, rather than requiring them to be
moved in the Makefile or packaging scripts.

Also, LINUX_MODULES=yes has a slightly different interpretation now.  If the
current target enables modules by default, then they will be built or else the
build fails.  Set it to no to disable modules entirely.  Modules are enabled
by default on all targets except armhf.  To enable modules on target that does
not enable them by default, set LINUX_MODULES to 'force'.  This allows
packaging configurations to avoid trying to set modules correctly for each
target; the defaults are set in one place, in the SConstruct file.  And it
means that a build fails if modules are enabled but cannot be built, rather
than silently being ignored if the kernel source directory is not found.

The init.d scripts for emerald and vortex have moved to their respective
kernel source directories and are installed from there, rather than being
duplicated or linked in the debian/ directory.  Since they seem to belong with
the modules packages, move them there and out of nidas-daq.  nidas-daq is now
just the init.d/dsm script, so as soon as we know no one uses that anymore, it
can be removed.