remove obsolete scripts
I think it's safe to say chroots are dead, at least as far as nidas builds are
concerned.  build_nidas.sh is not needed because the SCons tools adjust
PKG_CONFIG_PATH.  test_env is obsolete because the variant dir is no longer
just 'build', and setup_nidas.sh can be used instead to put a variant tree on
the path.