fix the merge and restore vortex package installs
in the refactor-builds branches, codename was derived automatically for
Raspbian jessie and buster, but for vortex it was set with -I on the command
line.  the automatic derivation was broken because VERSION on Ubuntu was two
words in upper case (Bionic Beaver).  now only the command-line setting is
used, but the code which derives the codename automatically has been fixed and
remains in case it could be useful.

also echo all the flock and reprepro commands, to make it easier to see how
they are being called.

skip gpg setup if a repo has no SignWith setting, such as when trying to test
the reprepro calls on a test repository.

lastly, no need to rm build* to trim the source archive as was done on the
branch, because master added -Ibuild*.