fix centos8 container builds
The official centos8 docker image no longer updates, so start from a community
image which already has the repository configs updated to pull from
vault.centos.org.  Install latest scons with pip.  Remove eol_scons as a build
requirement for RPM, so the requirement can be met with a local checkout and
not just with a package.  Also switch to building xmlrpc package in the
container.

Resolve duplicated comments and code in nidas.spec.  It looks like the libs
postun step was removing the custom file context with semanage, but then
adding it right back again and running restorecon.  I'm pretty sure that's not
what was intended.