copy nidas_udp_relay support files
since nidas_udp_relay is installed with nidas, include examples of the files
used to run it in production.  the plan is to replace the check script and
crontab with a systemd user unit and install that unit as part of nidas.
nidas_udp_relay will be modified to reconstruct the header from the xml file
rather than reading a separate file from the project configuration.  then the
nidas_eol_relay subdirectory can be removed from isfs/scripts because there
will no longer be any project-specific configuration required.