use lib instead of %{_lib}
On redhat lib64 systems, _lib is set to lib64, and nidas.spec used that in
nidas library paths to match the system directory name.  However, since nidas
libraries are not installed into the system library directory, the lib64
distinction is not really useful, so all the _lib references have been
replaced with 'lib'.  This is unrelated to _libdir, the full system library
path under which the pkg-config config file goes