fix 99-nidas.rules and remove nidas-daq
add 99-nidas.rules to nidas-modules-vortex package on debian, which should
restore missing file on vortex dsms; debian automatically installs package
rules files into /lib/udev/rules.d, but SConscript has been changed to do that
also, for both debian and redhat.

move the dsm.init script into nidas/scripts and install into bin, where
systems can copy or link it or call it as needed, but it's on it's way to
being deprecated and replaced with systemd units.  it was also the last file
still in the nidas-daq package, so now the nidas-daq package can be removed
also.

since the udev rules file only contains rules for nidas modules now, and it is
only installed as part of a modules pacakge, do not install the rules file
unless linux modules are enabled.