vortex container build fixes
use dolocal=yes in cnidas, to get uio48-dev build dependency for nidas, but do
not install nidas-dev in the container, since it doesn't seem to be needed for
a nidas build container, as expected.  add apt-get update to container build
steps, so apt is always updated before running apt-get install.