compute variance separately
the flux and covariance computations apparently share the same numsamples
count, but that is not the case for single-variable variance, where each
variable may have a different number of samples because of missing values.
this fixes the variance stats for the SOS snow pillows, where stats that
should have been missing values during periods of no data had zeros instead.