fix bug in min/max stats from unallocated nsamples
the min/max stats were using nsamples to test if the min/max was valid or not,
since if there were no samples then it should be nan.  however, _nSamples was
allocated with _nsum slots, and setupMinMax() was setting _nsum to zero.
Jacquie found this bug by noticing that the SOS flowcapt maxes occasionally
dropped to -1e37, because the test for nsamples was reading invalid memory and
sometimes returning non-zero even when there were no samples.  valgrind
confirmed this also.

the algorithm is slightly different now.  the min/max get set to nan, then the
first valid value automatically gets assigned as the min/max, so there is no
need to seed the min/max with a type-dependent limit.