turn off duplicate with variant_dir
according to the latest scons documentation, variant builds should work with
duplicate set to False, and from a little bit of testing, that seems to be the
case for nidas, so turn it off.  this helps IDEs find the right source file
path from the compiler output, and it avoids recursive searches finding all
the source files in all the variant dirs.  hopefully this doesn't break
anything on other targets.