prepare version 1.2.2
replace the generated changelog in the spec file with a manual one that just
links to the version being packaged; if modules rpm package is built, then may
as well package the modules under /lib/modules, not just the ones in
/opt/nidas/modules.  that also avoids warnings from rpm about installed but
unpackaged files.

rpm packages can be built with either the build_rpm script in tree or the
script in eol_scons, but the package versioning will be different.  the goal
is to transition to packaging explicit source releases, built from one source
release tar file for both debian and rpm, and eventually the local build_rpm
should become obsolete.