Support XML includes in XMLConfigService
Set parser->setXercesDoXInclude(true) in XMLConfigService

When selecting nodes for given DSM name, don't exclude any <dsm>s in the
<dsmcatalog> (those with an ID attribute).

Use dsm->expandString() and not project->expandString() when filtering
by DSM name in XMLConfigWriter. dsm->expandString() will be able to expand $SITE
if that is used for the DSM name (ISFS projects).