Cleanup of findDSM() methods in Project and Site
Maintain a clean distinction between searching the configuation for a DSM
by name, and by address.

Minimize DNS lookups, which are slow, especially when testing a config with
many DSMs on a system where DNS or /etc/hosts are not configured for
all the DSM names.

Improve log messages related to XMLConfigService, and findDSM().