remove timeout for simulated sonic serial sensor
...on a suspicion that the timeout is causing the serial_sensor test to fail
on alma9.  the logs indicated the sonic was timing out, causing dsm to reopen
it and an extra open message to show up in the log, so the test was killing
the sensor_sim processes prematurely (thinking all had closed because all the
data had been written).  there is no guarantee that all of the sensor_sim
processes will run concurrently (interleaved) in the background, so it looks
like sometimes on the alma9 server they actually run almost consecutively, or
maybe in batches of threes.  if we really need to test multiple sensors
writing samples interleaved in real time, then probably sensor_sim needs to
simulate multiple sensors in one process to interleave their writes.