switch to snapshot rpm packages in Jenkins
restore prior practice of building rpms for every commit, until we decide that
it would be good to package actual source releases also.  this is also
consistent with the debian packaging.  snapshot versions add commits since the
latest version tag and the commit hash, in a form which complies with rpm
conventions, without modifying the package build (release) number.