add more synchronization to sensor_sim for tests
use the close(open) trick on the pts to set HUP, so sensor_sim can wait until
the pts has been opened (HUP is cleared) before writing data.  this is easier
to synchronize and more accurate then sending the CONT signal.  add a timeout
which uses alarm() to guarantee that the sensor_sim process will exit, in case
something it is waiting for never happens.