merge run_test_2.sh into run_test.sh
these scripts were mostly the same, so this is an attempt to share shell
functions between the two tests and simplify the code.  since the test with
dsm_server does everything the other test did, just run the one with the
server by default.  the -i is no longer necessary, because the environment can
be setup to use the files under the variant_dir.