attempt to fix recent test failures on ubuntu32
add a sleep between sims finishing and killing processes, on a suspicion the
dsm and dsm_server processes are not being given time to read their final
samples and write them to the archive.  also disable sorting in the server
archive, since that is current and best practice, and it means the dsm and
dsm_server archive files should turn out identical.