Fix issues related to channel sampling sequence
The previous version used the variable index for the coefficients
of the polynomial A2D correction and should have used the channel index.
This would apply the wrong correction if the coefficients
varied by channel, and the channel number was not equal to the
variable index.

Skipping channels was supported, but the assignment of values
to variables in the output sample would be wrong if the
sequence of channels by variable index was out of order, e.g.
the first variable was channel 2 and second was channel 0.
Perhaps this never happened in configurations.

Use sscanf instead of strstr and atoi when parsing config samples.