Update A2Ds to use new A2DConverter classes
DSC_A2DSensor now supports reading a calibration file to
compute calibrated voltages using a 3rd order polynomial.

As before, DSMAnalogSensor uses a linear calibration correction,
and expects the calibration file to have 2 coefficients for each
channel.

Also, as before, A2D_Serial uses a 3rd order polynomial
calibration correction, expecting the calibration file to
have 4 coefficients for each chanel.

All classes support an "outputmode" <parameter> for the <sensor>,
which can be set in the XML to "counts", "volts", or "engineering".
Previously A2D_Serial expected "outputmode" to be a <parameter>
of a <sample>. Since it applies to all processed samples of the
sensor, it is now expected to be a parameter of the <sensor>