put all variants under build/
rather than pollute the src/ directory with several build_ variants, put them
all under the one build/ directory, as I should have done in the first place.
also setup the runtime PATH and LD_LIBRARY_PATH in this tool, so they are
available for all of the tests without setting that up in each one.  and
consolidate the settings to build against the variant also.  It should not be
a problem to add the runtime settings for the entire build, since nothing
besides nidas targets will be in those paths, but if it is, then those
settings can be in a different tool that only tests use.