extend DOMable base class with attribute handling
add methods to DOMable to retrieve attribute values and convert to different
types, to start to reduce the code duplication in fromDOMElement()
implementations.  also start on a mechanism to detect when an element node
attribute is not handled by any of the classes in the fromDOMElement()
implementation chain, so classes do not need to keep track of what attributes
are handled by the other classes.