isolate gsl dependency in implementation
no need to make the api itself dependent on whether GSL is available, so hide
the gsl in the implementation using the pimpl idiom.  getTransducerRotation()
and transducerShadowCorrection() are no-ops if GSL is not available.  also
COMPUTE_ABC2UVW_INVERSE has been defined for many years and many projects now,
so I don't think we need to keep both implementations in the code.  at this
point it might be fine to require GSL on all nidas platforms, but for now the
dependency is at least concentrated in one source file.