remove constParameters and constVariables
avoid keeping two containers around and in sync just so one can be returned as
a reference to const pointers.  instead return a copy of the container and let
the compiler optimize it.  that in turn required changes to iterators, so
simplify those also to avoid the need for explicit copy and assignment
methods.