move typed storage into base Parameter class
it does not add much overhead to keep all storage types in the base,
especially when the type is already being dispatched by case statements
anyway, and that moves towards not requiring the typed subclasses.  after
adding tests for the setValues() and getValues() methods, I discovered those
are not actually called anywhere in nidas, so they're ifdef'd out at the
moment.

type conversion is done in overloaded setValue() methods, but still needs to
be done for the get methods.