remove superfluous files
near as I can tell, the config, data, and scripts in this directory have been
completely superceded by the serial_sensor tests, and this directory has never
been included in the automated tests anyway.  all that remains is the
client.cc code, so may as well allow that to be built as a test.