flush the pipeline after all inputs disconnect
if all the raw inputs have disconnected, ie there are no more dsm
processes feeding data, that seems like a good time to flush any
samples remaining in the pipeline, rather than waiting for more
input or for a shutdown to flush the rest.  this and some extra
logging are attempts to resolve the vortex test failures, though the
test also needs to be fixed on fedora40 for some reason.