only call calcStatistics() on opened sensors
otherwise, sensor classes which read from their device (calling readBuffer()) trigger
race condition errors in helgring, because both the SensorOpener and SensorHandler
threads are accessin the same stats members.  while a sensor belongs to the
SensorOpener it should only be accessed by the SensorOpener thread.  and anyway,
it shouldn't hurt to not update statistics for a sensor which is not open yet.