tweak static initialization order for logger mutex
since it is harder to define when a module-scope static mutex is intialized,
replace it with a function static.  then add an init() method and call it from
the logging macro, to make sure the mutex exists (and is pushed onto the
finalization stack) before any LogContext.  otherwise if the mutex is destroyed
before the last LogContext, then the LogContext destructor accesses an
invalid mutex.  for some reason that has only been seen on ubuntu32 gcc 7.5.