add time and input criteria to SampleMatcher
significant refactor of SampleMatcher: RangeMatcher is now an exposed class
which is easier to create, use, and test in code, and it contains a filename
subtring and a time range which can be used to select or exclude samples,
though nothing uses those yet, since the sample id cache will have to be
rethought.  the additional criteria have been added to the samples specifier
syntax, one field for the time range and one for the input name.  also, rather
than return a bool indicating if criteria could be parsed, just throw an
exception so the message can give the user some idea of what couldn't be
parsed.  lots of testing updates.