throw EOF when FileSet has no files in time range
if a FileSet with a pattern does not match any files in the given time range,
that seems to me more like an empty file stream which should raise EOF, rather
than an IO error.  not being able to open an explicitly named file is still an
IO error.  this behaves more nicely with nidsmerge, where it's perfectly
natural for an input stream to not have any files in a specific time range.