do not add 1 second before computing next file time
the extra 1 second seems unnecessary: we're using integer math so the next
file time will always be on the next interval of the file length.  otherwise
it's possible to create a file within 1 second of the next interval which then
contains the next interval, and that is decidedly not what anyone expects when
writing data files at fixed intervals.  also some whitespace cleanup.