consolidate and upate some documentation
Update BUILDING.md with more current information, replacing the instructions
that are on the wiki and also in doxygen from nidas.h.  nidas.h instead now
points to the top README and gives an introduction to the namespaces.
PACKAGES.md gives an overview of packaging, both installing from and building,
replacing the outdated pages on the wiki.  README.md has been updated to point
to all the new locations for documentation.