Update parseBESTPOS() to handle new set of variables.
Novatel supports a GGA (position) packet with more decimal places for lat and lon, called GGALONG.  I'd like to switch to that for our higher rate data, and just record the BESTPOS at 1hz, just pulling out std deviations and extended number satellite data.

The rationale for this is that we are moving to 50hz, and the GGALONG is shorter than the BESTPOS.  Across RS232, this will matter.

50Hz BESTVEL and NMEA GGALONG.
1hz BESTPOS and NMEA RMC.