refactor utime tests
Add UTime constructor which takes integer seconds and microseconds, mostly so
tests can precisely set microseconds without using floating point.  Break down
utime unit tests further, and add more tests for format(), especially to help
isolate the test failures with %nf.  Remove the randomness in the testing near
the epoch and just explicitly test what should be the corner cases.  Also, it
turns out the integer math for the near-epoch test was overflowing, so the
times which we thought were being tested were not, so use long long for the
seconds so multiplying by USECS_PER_SEC doesn't overflow.

This commit still has test failures which need to be fixed.