fix tests for the change in formatted times
Separate the baseline test output into a baseline subdirectory, uncompressed,
to make it easier to diff and update files manually.  For now, most of the
diffs are working by removing the last two digits of the fractional seconds
with the sed_times function added to nidas_tests.sh, but eventually the
baseline files should be updated.