restore data_dump compare with less precision
i missed that somewhere along the way the data_dump test needed to allow a
second comparison with less precision, but the second comparison was not being
used anymore.  so remove the obsolete second comparison (diff_warn) to
simplify the test script, but use 4 digits of precision to restore the earlier
allowance for floating point differences, in case it's needed again.