fix override of virtual convert() method
I'm not sure if this was ever used or ever worked, but gcc 15 on fedora 42
now warns about the float convert() hiding the virtual double convert() in
VariableConverter, which I think means the convert() method called through a
base class pointer would never have used the float implementation in the
RebsLinear subclass.  Use the double signature and override to actually
override the base class virtual method.