clarify logic in SampleTag::getSite()
the previous implementation looked odd because it gave precedence to
getting the site from the dsm rather than the sensor.  in practice probably
the sensor always has the site set, and if it doesn't then the sensor
checks its own dsm for a site, so there's no reason for this method to
check the dsm.