copy data_stats.cc from master...
...and tweak til it builds on buster.  mostly this meant copying over UTime
API enhancements used by data_stats.  rather than also try to copy over
updates to Logger and NidasApp (for Precision argument), revert those
changes to data_stats since they are not neeeded to deploy on bookworm.

this branch is not intended to ever be merged back to anywhere, it is
strictly for deploying data_stats on bookworm arm64 for testing.  the
data_stats changes will eventually be deployed on the dsm by merging master
and buster.